requirements.txt
alembic==1.14.0
captum==0.7.0
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.1
-e git+ssh://git@github.com/Adaha-png/master_thesis@8da00ddc4fc9c82a817b68141dc21d6c6f30456e#egg=counterfactuals_shapley&subdirectory=package
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
greenlet==3.1.1
gymnasium==1.0.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.43.0
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
nsga-2==1.2.0
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pettingzoo==1.24.3
pillow==11.1.0
pygame==2.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
shap==0.46.0
six==1.17.0
slicer==0.0.8
SQLAlchemy==2.0.37
stable_baselines3==2.4.1
SuperSuit==3.9.3
sympy==1.13.1
threadpoolctl==3.5.0
tinyscaler==1.2.8
torch==2.5.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
